feat(oss):添加了阿里云OSS的对象存储服务,同时为用户信息添加了头像字段,头像存储在OSS

+++ src/main/java/pers/ly/mall/common/context/CurrentContext.java
@@ -1,13 +1,13 @@
 package pers.ly.mall.common.context;
 
 public class CurrentContext {
-    private static ThreadLocal<Integer> userIdLocal = new ThreadLocal<>();
+    private static ThreadLocal<Long> userIdLocal = new ThreadLocal<>();
 
-    public static void setUserId(int userId) {
+    public static void setUserId(Long userId) {
         userIdLocal.set(userId);
     }
 
-    public static Integer getUserId() {
+    public static Long getUserId() {
         return userIdLocal.get();
     }
